Added Combat Page
Added a new file for the combat page. Currently has little functionality. You have to access it by changing the hook in the inspector to 'Combat' instead of 'Item'. Also moved images into their own folder.
